pyproject.toml
[project]
name = "evaluate-llm-on-korean-dataset"
version = "0.1.0"
description = "Korean language proficiency evaluation for LLM/SLM models using Korean evaluation datasets"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-inference>=1.0.0b9",
    "boto3>=1.42.2",
    "datasets>=4.4.1",
    "huggingface-hub>=0.36.0",
    "ipykernel>=7.1.0",
    "langchain>=1.1.0",
    "langchain-aws>=1.1.0",
    "langchain-azure-ai>=0.1.8",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.1.0",
    "langchain-openai>=1.1.0",
    "numpy>=1.26.4",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "regex>=2025.9.18",
    "requests>=2.32.5",
    "safetensors>=0.7.0",
    "tabulate>=0.9.0",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
]
